// src/api/generated/api.ts
export interface ArticleFlat {
    'id'?: number;
    'title': string;
    'description'?: string;
    'local'?: string;
    'published_at'?: string;
    'revelance'?: number;
    'thumbnail_id'?: number;
    'links'?: Array<LinkFlat>;
    'skills'?: Array<SkillFlat>;
}
// ...
export interface AttachmentResp {
    'id': number;
    'title': string;
    'link': string;
    'description'?: string;
}
// ...
export interface LinkFlat {
    'id'?: number;
    'title'?: string;
    'link': string;
    'revelance'?: number;
    'description'?: string;
}
// ...
export interface SkillFlat {
    'id': number;
    'title': string;
    'revelance'?: number;
    'description'?: string;
}